pyproject.toml
[project]
name = "agentic-customer-support-olap"
version = "0.1.0"
description = "Multi-Agent OLAP benchmark on TPC-H via LLM-driven analytical queries"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "sqlalchemy>=2.0",
    "pymysql>=1.1",
    "pyyaml>=6.0",
    "cryptography",
    "backoff>=2.0",
    "httpx[socks]",
]

[tool.setuptools.packages.find]
include = ["olap_agent*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
